[Tizen] Change some matrix mediump to highp for image visual
when rendering FHD image, 1px at the end has a problem
only muse device has this problem

+++ dali-toolkit/internal/visuals/image-visual-shader-factory.cpp
@@ -38,9 +38,9 @@ const Vector4 FULL_TEXTURE_RECT(0.f, 0.f, 1.f, 1.f);
 
 const char* VERTEX_SHADER = DALI_COMPOSE_SHADER(
   attribute mediump vec2 aPosition;\n
-  uniform mediump mat4 uModelMatrix;\n
-  uniform mediump mat4 uViewMatrix;\n
-  uniform mediump mat4 uProjection;\n
+  uniform highp mat4 uModelMatrix;\n
+  uniform highp mat4 uViewMatrix;\n
+  uniform highp mat4 uProjection;\n
   uniform mediump vec3 uSize;\n
   uniform mediump vec4 pixelArea;
   varying mediump vec2 vTexCoord;\n
